Security Update
crypto-js is a library that provides cryptographic functionality to JavaScript programs. It includes functions for hashing, encryption and decryption, and message authentication. The version in the code you provided is version 4.1.1.

debug is a small library that provides simple debugging functionality for Node.js applications. It allows developers to output messages to the console with a specified namespace, so that only messages from certain parts of the code are displayed. The version in the code you provided is version 2.6.9.

delegates is a small library that allows for easy delegation of methods or properties from one object to another. This can be useful in implementing a prototype chain or for creating objects that delegate to other objects. The version in the code you provided is version 1.0.0.

depd is a library for handling deprecation warnings in Node.js applications. It allows developers to specify when a certain method or feature will be deprecated and will emit a deprecation warning when it is used. The version in the code you provided is version 2.0.0.

destroy is a small utility library that allows you to close and destroy streams in Node.js. It provides a consistent API for working with different types of streams, making it easier to clean up resources. The version in the code you provided is version 1.2.0.

detect-libc is a library that can be used to detect the underlying libc implementation on a given machine. libc is the C library that provides the system call interface on UNIX-like operating systems. The version in the code you provided is version 2.0.1.

dotenv is a library that allows you to load environment variables from a .env file in your project. This can be useful for storing sensitive configuration information that you don't want to hardcode into your application or check into version control. The version in the code you provided is version 16.0.3.